pyproject.toml
[tool.poetry]
name = "testfns"
version = "0.1.0"
description = ""
authors = ["Rezin.PV <p.rezin@limecreditgroup.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
python-dotenv = "^0.19.2"
pydantic = "^1.8.2"
django-ninja = "^0.16.1"
Django = "^4.0"
django-split-settings = "^1.1.0"
autoconfig = "^5.0.0"
ipython = "^7.30.1"
gunicorn = "^20.1.0"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
